langchain/requirements.txt
langchain
langchain-community
langchainhub
langchain-openai
langchain-chroma
bs4
python-dotenv
